[Refactor] Move ConnectorPrimitive files into connector_primitive module
