+++ be/CMakeLists.txt
@@ -963,13 +963,14 @@ ENDFUNCTION()
 add_subdirectory(${SRC_DIR}/agent)
 add_subdirectory(${SRC_DIR}/base)
 add_subdirectory(${SRC_DIR}/common)
-add_subdirectory(${SRC_DIR}/connector)
 add_subdirectory(${SRC_DIR}/column)
 add_subdirectory(${SRC_DIR}/formats)
 add_subdirectory(${SRC_DIR}/fs)
 add_subdirectory(${SRC_DIR}/fs_ext)
-add_subdirectory(${SRC_DIR}/module)
 add_subdirectory(${SRC_DIR}/exec_primitive)
+add_subdirectory(${SRC_DIR}/connector_primitive)
+add_subdirectory(${SRC_DIR}/connector)
+add_subdirectory(${SRC_DIR}/module)
 add_subdirectory(${SRC_DIR}/exec)
 add_subdirectory(${SRC_DIR}/exprs)
 add_subdirectory(${SRC_DIR}/gen_cpp)
