+++ be/module_boundary_manifest.json
@@ -772,14 +772,10 @@
       "doc_label": "ConnectorPrimitive",
       "summary": "Read-side connector contracts, DataSource, and DataSourceProvider default mechanics without concrete connectors, sinks, registry composition, storage, service, or full Exec coupling.",
       "owned_targets": ["ConnectorPrimitive"],
+      "owned_roots": ["be/src/connector_primitive"],
       "allowed_include_prefixes": [
-        "connector/connector.h",
-        "connector/data_source.h",
-        "connector/data_source_provider.h",
-        "exec_primitive/pipeline/scan/scan_morsel.h",
-        "exec_primitive/pipeline/scan/morsel_queue_builder.h",
-        "exec_primitive/pipeline/scan/dynamic_morsel_queue_builder.h",
-        "exec_primitive/runtime_filter/",
+        "connector_primitive/",
+        "exec_primitive/",
         "exprs/",
         "runtime/",
         "column/",
@@ -808,9 +804,7 @@
       "owned_targets": ["ConnectorBenchmark"],
       "allowed_include_prefixes": [
         "connector/benchmark/",
-        "connector/connector.h",
-        "connector/data_source.h",
-        "connector/data_source_provider.h",
+        "connector_primitive/",
         "exprs/",
         "runtime/",
         "column/",
@@ -836,9 +830,7 @@
       "owned_targets": ["ConnectorCacheStats"],
       "allowed_include_prefixes": [
         "connector/cache_stats/",
-        "connector/connector.h",
-        "connector/data_source.h",
-        "connector/data_source_provider.h",
+        "connector_primitive/",
         "storage/",
         "exprs/",
         "runtime/",
@@ -867,9 +859,7 @@
       "owned_targets": ["ConnectorJDBC"],
       "allowed_include_prefixes": [
         "connector/jdbc/",
-        "connector/connector.h",
-        "connector/data_source.h",
-        "connector/data_source_provider.h",
+        "connector_primitive/",
         "platform/",
         "fs/",
         "exprs/",
@@ -899,9 +889,7 @@
       "owned_targets": ["ConnectorElasticsearch"],
       "allowed_include_prefixes": [
         "connector/elasticsearch/",
-        "connector/connector.h",
-        "connector/data_source.h",
-        "connector/data_source_provider.h",
+        "connector_primitive/",
         "platform/",
         "exprs/",
         "runtime/",
@@ -930,9 +918,7 @@
       "owned_targets": ["ConnectorMySQL"],
       "allowed_include_prefixes": [
         "connector/mysql/",
-        "connector/connector.h",
-        "connector/data_source.h",
-        "connector/data_source_provider.h",
+        "connector_primitive/",
         "exprs/",
         "runtime/",
         "column/",
@@ -983,7 +969,7 @@
         "connector/jdbc/",
         "connector/elasticsearch/",
         "connector/mysql/",
-        "connector/connector.h",
+        "connector_primitive/",
         "connector/connector_registry.h",
         "fs/",
         "fs_ext/hdfs/",
@@ -1070,7 +1056,7 @@
       "allowed_include_prefixes": [
         "orchestration/",
         "data_workflows/",
-        "connector/data_source_provider.h",
+        "connector_primitive/",
         "column/",
         "exec/",
         "exec_primitive/",
