+++ be/src/connector/benchmark/AGENTS.md
@@ -8,7 +8,7 @@ Run `python3 build-support/check_be_module_boundaries.py --mode full` to validat
 ### ConnectorBenchmark (`connectorbenchmark`)
 Benchgen-backed benchmark connector implementation above connector contracts without registry composition, storage, service, or full Exec coupling.
 - Targets: `ConnectorBenchmark`
-- Allowed internal include prefixes: `connector/benchmark/`, `connector/connector.h`, `connector/data_source.h`, `connector/data_source_provider.h`, `exprs/`, `runtime/`, `column/`, `types/`, `common/`, `base/`, `gutil/`, `gen_cpp/`
+- Allowed internal include prefixes: `connector/benchmark/`, `connector_primitive/`, `exprs/`, `runtime/`, `column/`, `types/`, `common/`, `base/`, `gutil/`, `gen_cpp/`
 - Allowed target deps: `ConnectorPrimitive`, `Expr`, `Runtime`, `ChunkCore`, `ColumnCore`, `Types`, `Common`, `Base`, `Gutil`, `StarRocksGen`
 - Remediation: Keep ConnectorBenchmark limited to the benchgen connector implementation; move registration into ModuleBootstrap and avoid pulling Connector, storage, service, or full Exec code into the connector library.
 <!-- END GENERATED: BE MODULE HARNESSES -->
