+++ be/src/connector/file_connector.h
@@ -15,8 +15,8 @@
 #pragma once
 
 #include "column/vectorized_fwd.h"
-#include "connector/connector.h"
 #include "connector_chunk_sink.h"
+#include "connector_primitive/connector.h"
 #include "exec/file_scanner/file_scanner.h"
 
 namespace starrocks::connector {
