+++ be/src/connector/iceberg_connector.h
@@ -15,7 +15,7 @@
 #pragma once
 
 #include "column/vectorized_fwd.h"
-#include "connector/connector.h"
+#include "connector_primitive/connector.h"
 #include "exec/connector_scan_node.h"
 #include "exec/hdfs_scanner/hdfs_scanner.h"
