+++ be/src/connector/jdbc/AGENTS.md
@@ -8,7 +8,7 @@ Run `python3 build-support/check_be_module_boundaries.py --mode full` to validat
 ### ConnectorJDBC (`connectorjdbc`)
 JDBC connector implementation, scanner, driver manager, and type checker above connector contracts without registry composition, storage, service, or full Exec coupling.
 - Targets: `ConnectorJDBC`
-- Allowed internal include prefixes: `connector/jdbc/`, `connector/connector.h`, `connector/data_source.h`, `connector/data_source_provider.h`, `platform/`, `fs/`, `exprs/`, `runtime/`, `column/`, `types/`, `common/`, `base/`, `gutil/`, `gen_cpp/`
+- Allowed internal include prefixes: `connector/jdbc/`, `connector_primitive/`, `platform/`, `fs/`, `exprs/`, `runtime/`, `column/`, `types/`, `common/`, `base/`, `gutil/`, `gen_cpp/`
 - Allowed target deps: `ConnectorPrimitive`, `Platform`, `FileSystem`, `Expr`, `Runtime`, `ChunkCore`, `ColumnCore`, `Types`, `Common`, `Base`, `Gutil`, `StarRocksGen`
 - Core tests: `connector_jdbc_test`
 - Remediation: Keep ConnectorJDBC limited to JDBC read-side scan, driver management, and type checking; move registration into ModuleBootstrap and avoid pulling Connector, storage, service, or full Exec code into the connector library.
