+++ be/test/CMakeLists.txt
@@ -219,6 +219,7 @@ add_subdirectory(common)
 add_subdirectory(column)
 add_subdirectory(types)
 add_subdirectory(exec_primitive)
+add_subdirectory(connector_primitive)
 add_subdirectory(exec)
 add_subdirectory(connector)
 add_subdirectory(formats)
