+++ be/test/connector/CMakeLists.txt
@@ -17,31 +17,6 @@ if (APPLE)
     list(APPEND BE_TEST_MACOS_LINK_STUBS ../base/macos_link_stubs.cpp)
 endif()
 
-set(CONNECTOR_PRIMITIVE_TEST_LINK_LIBS
-    ConnectorPrimitive
-    ExecPrimitive
-    Expr
-    Runtime
-    ChunkCore
-    ColumnCore
-    Types
-    Common
-    Base
-    Gutil
-    StarRocksGen
-    ${STARROCKS_DEPENDENCIES}
-    ${STARROCKS_STATIC_LINK_LIBS}
-    ${WL_LINK_DYNAMIC} ${STARROCKS_SYSTEM_LINK_LIBS} ${WRAP_LINKER_FLAGS}
-    ${WL_START_GROUP}
-    gmock
-    gtest
-    ${WL_END_GROUP}
-)
-
-add_executable(connector_primitive_test ${BE_TEST_MACOS_LINK_STUBS} connector_primitive_test.cpp)
-target_link_libraries(connector_primitive_test ${CONNECTOR_PRIMITIVE_TEST_LINK_LIBS} gtest_main)
-set_target_properties(connector_primitive_test PROPERTIES COMPILE_FLAGS "-fno-access-control")
-
 set(CONNECTOR_CACHE_STATS_TEST_LINK_LIBS
     ConnectorCacheStats
     ${STARROCKS_DEPENDENCIES}
